refactor: replace DdbSingleTable with Table
The old DdbSingleTable class has been refactored into Table, which now
includes Table-specific methods from the client wrapper (e.g., table
instances can now call `describeTable` on themselves).
All table-specific types have also been moved out of src/types and into
the new src/Table dir.

BREAKING_CHANGE
