+++ src/Table/index.ts
@@ -1,0 +1,2 @@
+export * from "./Table";
+export * from "./types";
