--- src/types/tableTypes.ts
@@ -1,31 +1,0 @@
-/**
- * DynamoDB table properties set by the SDK
- */
-export type DdbSdkTableProperties = {
-  billingMode?: "PROVISIONED" | "PAY_PER_REQUEST";
-  provisionedThroughput?: {
-    read: number;
-    write: number;
-  };
-  // IDEA Add CreateTable params: "SSESpecification", "StreamSpecification", "TableClass".
-};
-
-/**
- * This type includes table properties set by the SDK, as well as configs that
- * are used by the DdbSingleTable class to control `ddb-single-table` behavior.
- */
-export type DdbTableConfigs = DdbSdkTableProperties & {
-  createIfNotExists: boolean;
-};
-
-/**
- * DynamoDB table indexes
- */
-export type DdbTableIndexes = {
-  [indexName: string]: {
-    name: string;
-    type: "GLOBAL" | "LOCAL";
-    indexPK: string;
-    indexSK?: string;
-  };
-};
